feat(cangjie_runtime): sync from upstream v1.2.0-beta.03

+++ runtime/src/Heap/Allocator/RegionManager.cpp
@@ -45,7 +45,7 @@ static size_t GetPageSize() noexcept
 #elif defined(__APPLE__)
     pageSize = static_cast<size_t>(sysconf(_SC_PAGESIZE));
 #else
-    pageSize = getpagesize();
+    pageSize = static_cast<size_t>(getpagesize());
 #endif
     return pageSize;
 }
